% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{Bs_update}
\alias{Bs_update}
\title{Update Spatially Varying Coefficients}
\usage{
Bs_update(
  Bs_old,
  x,
  z,
  site_id,
  off_set,
  neighbors,
  w,
  kse,
  theta,
  sigma2_beta,
  gamma,
  eta,
  corr_inv
)
}
\arguments{
\item{Bs_old}{Numeric matrix. Current values of Bs.}

\item{x}{Numeric matrix. Covariates for Bs.}

\item{z}{Numeric matrix. Covariates for fixed effects.}

\item{site_id}{Integer vector. Site/group ID for each observation.}

\item{off_set}{Numeric vector. Offset for each observation.}

\item{neighbors}{Numeric matrix. Adjacency matrix of site neighbors.}

\item{w}{Numeric vector. Polya-Gamma latent variables.}

\item{kse}{Numeric vector. Pseudo-residuals.}

\item{theta}{Numeric vector. Current fixed effect coefficients.}

\item{sigma2_beta}{Numeric. Prior variance for Bs.}

\item{gamma}{Numeric. Spatial correlation parameter.}

\item{eta}{Numeric vector. Spatial random effects.}

\item{corr_inv}{Matrix. Inverse of the temporal correlation matrix.}
}
\value{
A list with:
\item{Bs}{Updated spatially varying coefficients.}
\item{mu_beta}{Posterior mean of Bs.}
}
\description{
Samples the spatially varying coefficients (Bs) in the Gibbs sampler.
}
