% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{eta_update}
\alias{eta_update}
\title{Update Spatial Random Effects (Eta)}
\usage{
eta_update(
  Bs,
  x,
  z,
  site_id,
  off_set,
  neighbors,
  w,
  kse,
  theta,
  sigma2_eta,
  eta_old,
  rho
)
}
\arguments{
\item{Bs}{Numeric matrix. Current values of Bs.}

\item{x}{Numeric matrix. Covariates.}

\item{z}{Numeric matrix. Covariates for fixed effects.}

\item{site_id}{Integer vector. Site/group IDs.}

\item{off_set}{Numeric vector. Offset for each observation.}

\item{neighbors}{Numeric matrix. Adjacency matrix of site neighbors.}

\item{w}{Numeric vector. Polya-Gamma latent variables.}

\item{kse}{Numeric vector. Pseudo-residuals.}

\item{theta}{Numeric vector. Current fixed effect coefficients.}

\item{sigma2_eta}{Numeric. Prior variance for eta.}

\item{eta_old}{Numeric vector. Current values of eta.}

\item{rho}{Numeric. Spatial correlation parameter for eta.}
}
\value{
Numeric vector. Updated spatial random effects.
}
\description{
Samples the spatial random effects for each site in the Gibbs sampler.
}
