% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler for SVCGDLM}
\usage{
gibbs_sampler(
  mcmc_samples,
  y,
  x,
  z,
  site_id,
  neighbors,
  likelihood_indicator,
  off_set = NULL,
  theta = NULL,
  Bs_init = NULL,
  eta_init = NULL,
  sigma2_beta = NULL,
  gamma = NULL,
  phi = NULL,
  sigma2_eta = NULL,
  rho = NULL,
  log_r = NULL
)
}
\arguments{
\item{mcmc_samples}{Integer. Number of MCMC samples.}

\item{y}{Numeric vector. Outcome variable.}

\item{x}{Numeric matrix. Covariates for Bs.}

\item{z}{Numeric matrix. Covariates for fixed effects.}

\item{site_id}{Integer vector. Site/group IDs.}

\item{neighbors}{Numeric matrix. Adjacency matrix of site neighbors.}

\item{likelihood_indicator}{Integer. 0 for Bernoulli, 1 for Poisson/Negative Binomial.}

\item{off_set}{Numeric vector (optional). Offset.}

\item{theta}{Numeric vector. Initial fixed effect coefficients.}

\item{Bs_init}{Numeric matrix. Initial Bs.}

\item{eta_init}{Numeric vector. Initial eta.}

\item{sigma2_beta}{Numeric. Prior variance for Bs.}

\item{gamma}{Numeric. Spatial correlation parameter.}

\item{phi}{Numeric. Temporal correlation parameter.}

\item{sigma2_eta}{Numeric. Prior variance for eta.}

\item{rho}{Numeric. Spatial correlation parameter for eta.}

\item{log_r}{Numeric. Log of the r parameter (for likelihood_indicator = 1).}
}
\value{
A list with:
\item{mcmc_samples}{Number of samples.}
\item{Bs}{List of sampled Bs.}
\item{mu_beta}{List of sampled mu_beta.}
\item{eta}{List of sampled eta.}
\item{w}{List of sampled Polya-Gamma variables.}
}
\description{
Runs the Gibbs sampler for the Spatially Varying Coefficient Generalized Dynamic Linear Model.
}
