% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{log_pg_density_vectorized}
\alias{log_pg_density_vectorized}
\title{Vectorized Log Density for Polya-Gamma Distribution}
\usage{
log_pg_density_vectorized(x, b, truncation = 100)
}
\arguments{
\item{x}{Numeric vector. Values at which to evaluate the density.}

\item{b}{Numeric vector. Polya-Gamma parameters.}

\item{truncation}{Integer. Number of terms in the series expansion.}
}
\value{
Numeric vector. Log density values.
}
\description{
Computes the log density of the Polya-Gamma distribution for vectors of x and b.
}
