% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{optimize_model}
\alias{optimize_model}
\title{Optimize SVCGDLM Parameters via MCEM}
\usage{
optimize_model(likelihood_indicator, par0, nll, cl)
}
\arguments{
\item{likelihood_indicator}{Integer. 0 for Bernoulli, 1 for Poisson/Negative Binomial.}

\item{par0}{Numeric vector. Initial parameter values.}

\item{nll}{Function. Negative log-likelihood function to be minimized.}

\item{cl}{Parallel cluster object.}
}
\value{
The output of \code{optimParallel}.
}
\description{
Optimizes model parameters in the MCEM algorithm using parallel L-BFGS-B.
}
