% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{polya_gamma_log_density}
\alias{polya_gamma_log_density}
\title{Log Density of the Polya-Gamma Distribution}
\usage{
polya_gamma_log_density(x, b = 1, likelihood_indicator = 0)
}
\arguments{
\item{x}{Numeric. Value(s) at which to evaluate the density.}

\item{b}{Numeric. Parameter for the Polya-Gamma distribution.}

\item{likelihood_indicator}{Integer. 0 for Bernoulli, 1 for Negative Binomial.}
}
\value{
Numeric. Log density value(s).
}
\description{
Computes the log density of the Polya-Gamma distribution using a finite series expansion.
}
