% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{precompute_pg_log_terms}
\alias{precompute_pg_log_terms}
\title{Precompute Terms for Polya-Gamma Log Density}
\usage{
precompute_pg_log_terms(n_terms = 200)
}
\arguments{
\item{n_terms}{Integer. Number of terms to precompute.}
}
\value{
A list of precomputed terms.
}
\description{
Precomputes log factorials and powers for faster evaluation of the Polya-Gamma log density.
}
