% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{Q_theta}
\alias{Q_theta}
\title{Compute Q Function for MCEM Algorithm}
\usage{
Q_theta(
  L = L,
  likelihood_indicator,
  param,
  omega_mean,
  Omega,
  Bs_mean,
  mu_beta_mean,
  eta_rep,
  chain_eta2,
  chain_Bs2,
  X_block,
  site_id,
  y,
  z,
  D
)
}
\arguments{
\item{L}{Integer. Number of time points.}

\item{likelihood_indicator}{Integer. 0 for Bernoulli, 1 for Poisson/Negative Binomial.}

\item{param}{Numeric vector. Model parameters.}

\item{omega_mean}{Numeric vector. Mean Polya-Gamma variables.}

\item{Omega}{Matrix. Block-diagonal matrix of omega.}

\item{Bs_mean}{Numeric matrix. Mean of Bs samples.}

\item{mu_beta_mean}{Numeric vector. Mean of mu_beta samples.}

\item{eta_rep}{Numeric vector. Replicated eta.}

\item{chain_eta2}{Numeric matrix. Eta samples.}

\item{chain_Bs2}{Numeric array. Bs samples.}

\item{X_block}{Matrix. Block-diagonal design matrix.}

\item{site_id}{Integer vector. Site/group IDs.}

\item{y}{Numeric vector. Outcome variable.}

\item{z}{Numeric matrix. Covariates for fixed effects.}

\item{D}{Matrix. Difference matrix for spatial structure.}
}
\value{
Numeric. The value of the Q function.
}
\description{
Computes the expected complete-data log-likelihood (Q function) in the MCEM algorithm for SVCGDLM.
}
