% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCGDLM.R
\name{temporal_corr_fun}
\alias{temporal_corr_fun}
\title{Temporal Correlation Matrix and Its Inverse}
\usage{
temporal_corr_fun(L, phi)
}
\arguments{
\item{L}{Integer. The number of time points (must be positive).}

\item{phi}{Numeric. The temporal decay parameter (must be positive).}
}
\value{
A list with:
\item{temporal_corr_inv}{Inverse of the temporal correlation matrix.}
\item{log_deter}{Log determinant of the temporal correlation matrix.}
}
\description{
Computes the temporal correlation matrix for a given length and decay parameter (phi), its inverse, and the log determinant.
}
\examples{
temporal_corr_fun(5, 0.1)

}
